+++ src/main/java/io/fabric8/maven/docker/access/hc/DockerAccessWithHcClient.java
@@ -161,6 +161,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
             request.addProperty("Detach", false);
             request.addProperty("Tty", true);
 
+            log.verbose(Logger.LogVerboseCategory.API,"POSTing to %s with %s", url, request);
             delegate.post(url, request.toString(), createExecResponseHandler(outputSpec), HTTP_OK);
         } catch (Exception e) {
             throw new DockerAccessException(e, "Unable to start container id [%s]", containerId);
@@ -202,6 +203,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
         request.add("Cmd", JsonFactory.newJsonArray(arguments.getExec()));
 
         String execJsonRequest = request.toString();
+        log.verbose(Logger.LogVerboseCategory.API,"POSTing to %s with %s", url, execJsonRequest);
         try {
             String response = delegate.post(url, execJsonRequest, new ApacheHttpClientDelegate.BodyResponseHandler(), HTTP_CREATED);
             JsonObject json = JsonFactory.newJsonObject(response);
@@ -225,6 +227,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
 
         try {
             String url = urlBuilder.createContainer(containerName);
+            log.verbose(Logger.LogVerboseCategory.API,"POSTing to %s with %s", url, createJson);
             String response =
                     delegate.post(url, createJson, new ApacheHttpClientDelegate.BodyResponseHandler(), HTTP_CREATED);
             JsonObject json = JsonFactory.newJsonObject(response);
@@ -262,6 +265,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
     public void buildImage(String image, File dockerArchive, BuildOptions options) throws DockerAccessException {
         try {
             String url = urlBuilder.buildImage(image, options);
+            log.verbose(Logger.LogVerboseCategory.API,"POSTing to %s with contents of file %s", url, dockerArchive);
             delegate.post(url, dockerArchive, createBuildResponseHandler(), HTTP_OK);
         } catch (IOException e) {
             throw new DockerAccessException(e, "Unable to build image [%s]", image);
@@ -273,6 +277,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
             throws DockerAccessException {
         try {
             String url = urlBuilder.copyArchive(containerId, targetPath);
+            log.verbose(Logger.LogVerboseCategory.API,"PUTing to %s with contents of file %s", url, archive);
             delegate.put(url, archive, HTTP_OK);
         } catch (IOException e) {
             throw new DockerAccessException(e, "Unable to copy archive %s to container [%s] with path %s",
@@ -394,6 +399,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
             throws DockerAccessException {
         try {
             String url = urlBuilder.removeContainer(containerId, removeVolumes);
+            log.verbose(Logger.LogVerboseCategory.API,"DELETEing %s", url);
             delegate.delete(url, HTTP_NO_CONTENT);
         } catch (IOException e) {
             throw new DockerAccessException(e, "Unable to remove container [%s]", containerId);
@@ -404,6 +410,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
     public void loadImage(String image, File tarArchive) throws DockerAccessException {
         String url = urlBuilder.loadImage();
 
+        log.verbose(Logger.LogVerboseCategory.API,"POSTing to %s with contents of file %s", url, tarArchive);
         try {
             delegate.post(url, tarArchive, new BodyAndStatusResponseHandler(), HTTP_OK);
         } catch (IOException e) {
@@ -531,6 +538,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
         log.debug("Network create config: " + createJson);
         try {
             String url = urlBuilder.createNetwork();
+            log.verbose(Logger.LogVerboseCategory.API,"POSTing to %s with %s", url, createJson);
             String response =
                     delegate.post(url, createJson, new ApacheHttpClientDelegate.BodyResponseHandler(), HTTP_CREATED);
             log.debug(response);
@@ -552,6 +560,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
             throws DockerAccessException {
         try {
             String url = urlBuilder.removeNetwork(networkId);
+            log.verbose(Logger.LogVerboseCategory.API,"DELETEing %s", url);
             int status = delegate.delete(url, HTTP_OK, HTTP_NO_CONTENT, HTTP_NOT_FOUND);
             return status == HTTP_OK || status == HTTP_NO_CONTENT;
         } catch (IOException e) {
@@ -569,6 +578,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
         try
         {
             String url = urlBuilder.createVolume();
+            log.verbose(Logger.LogVerboseCategory.API,"POSTing to %s with %s", url, createJson);
             String response =
                     delegate.post(url,
                                   createJson,
@@ -590,6 +600,7 @@ public class DockerAccessWithHcClient implements DockerAccess {
     public void removeVolume(String name) throws DockerAccessException {
         try {
             String url = urlBuilder.removeVolume(name);
+            log.verbose(Logger.LogVerboseCategory.API,"DELETEing %s", url);
             delegate.delete(url, HTTP_NO_CONTENT, HTTP_NOT_FOUND);
         } catch (IOException e) {
             throw new DockerAccessException(e, "Unable to remove volume [%s]", name);